update mh,mn to float

+++ database/migrations/2023_02_28_202750_create_tbt_statistic_months_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
 						$table->foreignId('tbt_statistic_id')->constrained('tbt_statistics')->onDelete('cascade');
 						$table->smallInteger('month_code');
 						$table->string('month');
-						$table->integer('manpower');
-						$table->integer('manhours');
+						$table->float('manpower');
+						$table->float('manhours');
             $table->timestamps();
         });
     }
